feat: add code editor in GrographyDrawer
